pyproject.toml
[project]
name = "whisper-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "redis>=7.4.0",
    "rq>=2.7.0",
    "uvicorn>=0.42.0",
]
